This is a big commit: - Changed the use of floats to doubles in the program as a way of increasing precision: this might prove excessive and, if we need more memory we can always fall back to float. - Bugfix: in the RibisimeSimulator class, we wrongly swapped a near7f by a nea4f: this is fixed in this commit. - In main.cpp: created procedure to: read the concentrations files, and, given a number of iterations, generate two files:   a) a file with times of decoding and translocating for each iteration (the equivalent of samplevectors.csv)   b) a file with the average decoding time of all codons and the average translocating time (codon 'tra').

+++ concentrations_reader.cpp
@@ -25,9 +25,9 @@ void concentrations_reader::load_concentrations(std::string file_name)
     contents.clear();
     std::string codon;
     std::string three_letter;
-    float wc_cognate_conc;
-    float wobblecognate_conc;
-    float nearcognate_conc;
+    double wc_cognate_conc;
+    double wobblecognate_conc;
+    double nearcognate_conc;
     std::string tmp_str;
     std::vector<std::string> stop_codons = {"UAG", "UAA", "UGA"}; // list of stop codons.
     bool header = true;
