+++ concentrations_reader.h
@@ -6,9 +6,9 @@ namespace csv_utils {
     struct concentration_entry{
         std::string codon;
         std::string three_letter;
-        float wc_cognate_conc;
-        float wobblecognate_conc;
-        float nearcognate_conc;
+        double wc_cognate_conc;
+        double wobblecognate_conc;
+        double nearcognate_conc;
     };
     
     class concentrations_reader
