+++ gillespie.h
@@ -21,9 +21,9 @@ namespace Simulations {
         int iteration_limit;
         Eigen::MatrixXi initial_populations;
         ReactionsSet reactions;
-        std::vector<float> dt_history;
+        std::vector<double> dt_history;
         std::vector<Eigen::MatrixXi> population_history;
-        float total_time;
+        double total_time;
         
     };
 }
